Since draw calls are sorted first by render trait layer and then by
depth of the owning game object in the scene tree, we must remove and reinsert the render trait into the draw order tree whenever its
associated game object is moved in the scene tree such that it has a new
depth. We do not have to resort the draw calls when the render layer
changes, because changing the render layer for a render-trait is not
possible.